// types/response_codes.ts
export enum successCodes {
	SUCCESS = "SUCCESS",
	AUTH_SUCCESS = "S_AUTHENTICATION",
	LOGOUT_SUCCESS = "S_LOGOUT",
	USER_CREATED = "S_USER_CREATED",
	USER_DELETED = "S_USER_DELETED",
	USER_LASTNAME_UPDATED = "S_LASTNAME_UPDATED",
	USER_PASSWORD_UPDATED = "S_PASSWORD_UPDATED",
	USER_USERNAME_UPDATED = "S_USERNAME_UPDATED",
	USER_EMAIL_UPDATED = "S_EMAIL_UPDATED",
	AVATAR_UPLOADED = "S_AVATAR_UPLOADED",
	AVATAR_DELETED = "S_AVATAR_DELETED",
}

export enum errorCodes {
	AVATAR_IS_NULL = "E_AVATAR_IS_NULL",
	INVALID_OLD_PASSWORD = "E_INVALID_OLD_PASSWORD",
	VALIDATION_ERROR = "E_VALIDATION_ERROR",
	UNAUTHORIZED = "E_UNAUTHORIZED_ACCESS",
	NOT_FOUND = "E_NOT_FOUND",
	SERVER_ERROR = "E_SERVER_ERROR",
	BAD_REQUEST = "E_BAD_REQUEST",
	INVALID_CREDENTIALS = "E_INVALID_CREDENTIALS",
}